packages/pandemic-core/pyproject.toml
[project]
name = "pandemic-core"
version = "0.0.1"
description = "Pandemic core daemon for edge computing"
readme = "README.md"
license = {text = "MIT"}
keywords = ["edge", "daemon", "systemd"]
authors = [{name = "Pandemic Team"}]
dependencies = [
    "pyyaml>=6.0",
    "psutil>=5.9.0",
    "pandemic-common==0.0.1"
]

[project.optional-dependencies]
test = ["pytest>=7.0", "pytest-asyncio>=0.21", "pytest-cov>=4.0"]
dev = ["black>=23.0", "isort>=5.0", "mypy>=1.0", "flake8>=6.0"]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
pandemic = "pandemic_core.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"